abstracted usfm method to helper
